requirements.txt
openai>=1.40.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
